pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Zhang Jian <thiszj@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^2.2"
djangorestframework = "^3.12.4"
djongo = "^1.3.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
